--> Golden master of v2.6.0: Awaiting final tests on hardware before release. --> Fixed issue with reading status register 2 and the Suspend operation where the library was not checking the suspend status correctly.
