+++ extras/Changes.log
@@ -4,22 +4,28 @@
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 //				 	 						Version 2.6.0 									   	   //
 //									Author: Prajwal Bhattaram									   //
-//					  						22.11.2016											   //
+//					  						15.04.2017											   //
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
-Enhancements:
---> Added a new error code. Library can now detect non-responsive chips - bad wiring or otherwise.
---> Streamlined variables to make code structure better.
---> RTL8195A compatibility tested and enabled by @boseji <salearj@hotmail.com> on 02.03.17. Code modified to fit with the library structure by Prajwal Bhattaram on 14.04.17.
---> Changed library structure to enable the addition of multi-flash compatibility.
---> Added support for the Simblee module.
---> Compatible with the ESP32 core for Arduino as of commit 67fd652 (https://github.com/espressif/arduino-esp32/tree/67fd6521777a16e406639a0eb0d113bf2e896f97). The ESP32 core currently does not support analogRead and so randomSeed(AnalogRead(A0)) cannot be used. Also, for some unknown reason, SPI clock speeds higher than board speed/4 are not stable and so, the clock speed for ESP32 dev boards has currently been throttled to 20MHz.
---> moved `#define RUNDIAGNOSTIC` & `#define HIGHSPEED` to SPIFlash.h from SPIFlash.cpp.
---> Now works with non Winbond modules (beta) by taking the flash memory size in bits as an argument in flash.begin('CHIPSIZE');
---> Formatted code to be better human readable.
---> Changed the internal _troubleshoot() function to use fewer resources.
+Bugs Squashed:
+--> Fixed issue with reading status register 2 and the Suspend operation where the library was not checking the suspend status correctly.
 
 Deletions:
 --> flash.getChipName() has been removed as it is superfluous now that the library is now on its way to multi-chip compatibility.
+--> Library is not compatible with the ATTiny85 anymore. Currently working on a fix which will hopefully be rolled out in v2.7.0
+
+New Boards supported:
+--> RTL8195A compatibility tested and enabled by @boseji <salearj@hotmail.com> on 02.03.17. Code modified to fit with the library structure by Prajwal Bhattaram on 14.04.17.
+--> Compatible with the ESP32 core for Arduino as of commit 67fd652 (https://github.com/espressif/arduino-esp32/tree/67fd6521777a16e406639a0eb0d113bf2e896f97). The ESP32 core currently does not support analogRead and so randomSeed(AnalogRead(A0)) cannot be used. Also, for some unknown reason, SPI clock speeds higher than board speed/4 are not stable and so, the clock speed for ESP32 dev boards has currently been throttled to 20MHz.
+--> Added support for the Simblee module.
+
+Enhancements:
+--> Added a new error code. Library can now detect non-responsive chips - bad wiring or otherwise.
+--> Streamlined variables to make code structure better.
+--> Changed library structure to enable the addition of multi-flash compatibility in the near future.
+--> moved `#define RUNDIAGNOSTIC` & `#define HIGHSPEED` to SPIFlash.h from SPIFlash.cpp.
+--> Now works with other Winbond modules (not in the official supported module list) (beta) by taking the flash memory size in bits as an argument in flash.begin('CHIPSIZE');
+--> Formatted code to be better human readable.
+--> Changed the internal _troubleshoot() function to use fewer resources.
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 //				 	 						Version 2.5.0 									   	   //
 //									Author: Prajwal Bhattaram									   //
