+++ library.properties
@@ -1,10 +1,10 @@
 name=SPIFlash
-version=2.5.0
+version=2.6.0
 author=Prajwal Bhattaram
 maintainer=Prajwal Bhattaram <marzogh@icloud.com>
 sentence=Winbond SPI flash library for Arduino.
 paragraph=This library enables read, write, erase and power functions on the following Winbond NOR Flash chips - W25X05CL, W25X10BV, W25X20BV, W25X40BV, W25Q80BV, W25Q16BV, W25Q32BV, W25Q64BV & W25Q128BV. Refer to change log for further information about this release.
 category=Data Storage
 url=https://github.com/Marzogh/SPIFlash
-architectures=*
+architectures=avr,sam,samd,esp8266,esp32,simblee
 includes=SPIFlash.h
